Update CMD Quicksheet for SummaXR Custom Painterly 1.20v1.0
SummaXR Custom Painterly 1.20v1.0 stripped out a lot of custom model textures which had gone unimplemented, since they relied on base item textures which were changing, and their implementation was pending further work on the server.  They will be re-added as needed when the server-side revisions are done, based on the new textures.